Rename config -> Config, logger -> Logger, server -> main_controller, CommandInterface -> command_interface, edit -> workspace
