+++ python-template/src/lib/command_interface.py
@@ -5,11 +5,11 @@ import signal
 import socket
 import sys
 
+from Config import EXIT_COMMAND, SOCK_HOST, SOCK_PORT, WORDLIST_FILE, log
 from DeviceList import AllowedDevicesNodeIDs
+from Logger import pprint
 
-from config import EXIT_COMMAND, SOCK_HOST, SOCK_PORT, WORDLIST_FILE, log
-from edit import encrypt
-from logger import pprint
+from workspace import encrypt
 
 device_list = dict()
 wordlist = list()
